app/requirements.txt
Flask
psycopg2-binary # For PostgreSQL connection